+++ src/components/Menu/Menu.js
@@ -1,6 +1,6 @@
 import { ProjectsMenu } from "./ProjectsMenu";
 import { WorkMenu } from "./WorkMenu";
-import "../style/index.css";
+import "../../style/index.css";
 export const Menu = (props) => {
   return (
     <div id="menu" className="reveal-left">
