+++ src/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import ReactDOM from "react-dom/client";
 import "./style/index.css";
 import App from "./App";
